Cambios Promociones y descuentos web services

+++ src/java/co/com/lavapp/modelo/dto/Costo_TO.java
@@ -4,6 +4,7 @@
  * and open the template in the editor.
  */
 package co.com.lavapp.modelo.dto;
+
 /**
  *
  * Objeto de negocios que modelo un Costo
@@ -24,7 +25,7 @@ public class Costo_TO {
      *
      * Columna valor
      */
-    private int valor;
+    private double valor;
 
     /**
      *
@@ -45,14 +46,14 @@ public class Costo_TO {
     public Costo_TO(int idCosto) {
         this.idCosto = idCosto;
     }
-        
-    public Costo_TO(int valor, SubProducto_TO subProducto, Zona_TO zona) {
+
+    public Costo_TO(double valor, SubProducto_TO subProducto, Zona_TO zona) {
         this.valor = valor;
         this.subProducto = subProducto;
         this.zona = zona;
     }
 
-    public Costo_TO(int idCosto, int valor, SubProducto_TO subProducto, Zona_TO zona) {
+    public Costo_TO(int idCosto, double valor, SubProducto_TO subProducto, Zona_TO zona) {
         this.idCosto = idCosto;
         this.valor = valor;
         this.subProducto = subProducto;
@@ -64,10 +65,6 @@ public class Costo_TO {
         this.valor = valor;
         this.zona = zona;
     }
-    
-    
-
-    
 
     //Getter and Setter
     public int getIdCosto() {
@@ -78,11 +75,11 @@ public class Costo_TO {
         this.idCosto = idCosto;
     }
 
-    public int getValor() {
+    public double getValor() {
         return valor;
     }
 
-    public void setValor(int valor) {
+    public void setValor(double valor) {
         this.valor = valor;
     }
 
@@ -93,7 +90,7 @@ public class Costo_TO {
     public void setSubProducto(SubProducto_TO subProducto) {
         this.subProducto = subProducto;
     }
-    
+
     public Zona_TO getZona() {
         return zona;
     }
@@ -101,7 +98,6 @@ public class Costo_TO {
     public void setZona(Zona_TO zona) {
         this.zona = zona;
     }
-    
 
     @Override
     public String toString() {
